Update to follow castwide's temp file approach
