+++ lib/solargraph/language_server/message/text_document/formatting.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 require 'rubocop'
-require 'shellwords'
-require 'solargraph'
+require 'securerandom'
 
 module Solargraph
   module LanguageServer
@@ -11,23 +10,29 @@ module Solargraph
         class Formatting < Base
           def process
             filename = uri_to_file(params['textDocument']['uri'])
-            filename = 'tmp.rb' if filename.nil? or filename.empty?
+            # Make the temp file in the original file's directory so RuboCop
+            # detects the correct configuration
+            tempfile = File.join(File.dirname(filename), "_tmp_#{SecureRandom.hex(8)}_#{File.basename(filename)}")
             original = host.read_text(params['textDocument']['uri'])
-            options, paths = RuboCop::Options.new.parse(['-a', '-f', 'fi', filename])
-            options[:stdin] = original
-            runner = RuboCop::Runner.new(options, RuboCop::ConfigStore.new)
-            result = Solargraph::Diagnostics::RubocopHelpers.redirect_stdout{ runner.run(paths) }
-            return format(original, result) if result && !result.empty?
-          rescue RuboCop::ValidationError, RuboCop::ConfigNotFoundError => e
-            set_error(Solargraph::LanguageServer::ErrorCodes::INTERNAL_ERROR, e)
+            File.write tempfile, original
+            begin
+              options, paths = RuboCop::Options.new.parse(['-a', '-f', 'fi', tempfile])
+              RuboCop::Runner.new(options, RuboCop::ConfigStore.new).run(paths)
+              result = File.read(tempfile)
+              File.unlink tempfile
+              format original, result
+            rescue RuboCop::ValidationError, RuboCop::ConfigNotFoundError => e
+              set_error(Solargraph::LanguageServer::ErrorCodes::INTERNAL_ERROR, "[#{e.class}] #{e.message}")
+              File.unlink tempfile
+            end
           end
 
+          private
+
+          # @param original [String]
+          # @param result [String]
+          # @return [void]
           def format original, result
-            lines = result.lines
-            index = lines.index{|l| l.start_with?('====================')}
-            formatted = lines[index+1..-1].join
-            # The response is required to send an explicit range. Text edits
-            # with null ranges get ignored. See castwide/vscode-solargraph#83
             if original.end_with?("\n")
               ending = {
                 line: original.lines.length,
@@ -49,7 +54,7 @@ module Solargraph
                     },
                     end: ending
                   },
-                  newText: formatted
+                  newText: result
                 }
               ]
             )
